[송성학] fix: 배포 준비를 위한 버그 픽스
